Editor: Collada Importer - Part 1 - Introduce igExtSplitterV - Refactor 3D views to Viewport3D - Move Model-related classes to seperate folder - Add Log control - Add --dx9 flag for forcing ANGLE backend

+++ src/Editor/LancerEdit/UtfTab.Popups.cs
@@ -302,7 +302,7 @@ namespace LancerEdit
         {
             ImGui.Text("Name: ");
             ImGui.SameLine();
-            bool entered = ImGui.InputText("", text.Pointer, text.Size, InputTextFlags.EnterReturnsTrue, text.Callback);
+            bool entered = ImGui.InputText("", text.Pointer, (uint)text.Size, InputTextFlags.EnterReturnsTrue, text.Callback);
             if (data.DoFocus) ImGui.SetKeyboardFocusHere();
             if (entered || ImGui.Button("Ok"))
             {
@@ -324,7 +324,7 @@ namespace LancerEdit
         {
             ImGui.Text("Name: ");
             ImGui.SameLine();
-            bool entered = ImGui.InputText("", text.Pointer, text.Size, InputTextFlags.EnterReturnsTrue, text.Callback);
+            bool entered = ImGui.InputText("", text.Pointer, (uint)text.Size, InputTextFlags.EnterReturnsTrue, text.Callback);
             if (data.DoFocus) ImGui.SetKeyboardFocusHere();
             if (entered || ImGui.Button("Ok"))
             {
